Add /relocate command
Currently the only way to relocate a message is to use the "Send to
Memes Channel" option in the context menu. This (a) only allows
relocating messages to the configured memes channel, and (b) doesn't
work on mobile.

This commit adds a new /relocate command which allows relocating
messages to any channel and works on mobile devices.