Optionally remove speech bubbles from relocated messages
Speech bubbles prevent upvotes and downvotes on messages in the memes
channel. After relocating messages was introduced, sometimes they were
used on memes posted in other channels to prevent upvotes or downvotes if
they were moved to the memes channel in the future.

This commit adds an option to circumvent this and post the meme the way
it should have been posted in the first place.