Create a custom bot class
This allows the database connection to be passed into extensions without
using global variables, since the database connection can be stored in
the bot object.

The constructor does not support all of the parameters of the base
class, because that would make the code unreadable and/or break static
type checking.