Move /relocate command to a cog
This commit also moves the Send to Memes Channel context menu action,
which does the same thing with some preset parameters.

Additional permission checks are added to ensure the user can manage
messages in both the source and destination channels, and that the bot
has sufficient permissions to perform the operation.